+++ system/src/Grav/Framework/Session/Session.php
@@ -9,6 +9,7 @@
 
 namespace Grav\Framework\Session;
 
+use Grav\Common\User\Interfaces\UserInterface;
 use Grav\Framework\Session\Exceptions\SessionException;
 
 /**
@@ -198,13 +199,12 @@ class Session implements SessionInterface
             throw new SessionException('Failed to start session: ' . $error, 500);
         }
 
-        try {
-            if ($user && !$user->isValid()) {
-                throw new \RuntimeException('Invalid User');
-            }
-        } catch (\TypeError|\RuntimeException $e) {
-            $this->clear();
-            throw new SessionException('Invalid User Object, restarting session', 500, $e);
+        $this->started = true;
+
+        if ($user && (!$user instanceof UserInterface || !$user->isValid())) {
+            $this->invalidate();
+
+            throw new SessionException('Invalid User object, session destroyed.', 500);
         }
 
         // Extend the lifetime of the session.
@@ -221,8 +221,6 @@ class Session implements SessionInterface
                 $params['httponly']
             );
         }
-
-        $this->started = true;
 
         return $this;
     }