+++ CHANGELOG.md
@@ -238,7 +238,10 @@
 1. [](#bugfix)
     * Fixed incorrect routing caused by `str_replace()` in `Uri::init()` [#2754](https://github.com/getgrav/grav/issues/2754)
     * Fixed session cookie is being set twice in the HTTP header [#2745](https://github.com/getgrav/grav/issues/2745)
+    * Fixed session not restarting if user was invalid (downgrading from Grav 1.7)
     * Fixed filesystem iterator calls with non-existing folders
+    * Fixed `checkbox` field not being saved, requires also Form v4.0.2 [#1225](https://github.com/getgrav/grav/issues/1225)
+    * Fixed `validation: strict` not working in blueprints [#1273](https://github.com/getgrav/grav/issues/1273)
 
 # v1.6.19
 ## 12/04/2019
@@ -248,7 +251,7 @@
 1. [](#bugfix)
     * Fixed fatal error when calling `{{ grav.undefined }}`
     * Fixed multiple issues when there are no pages in the site
-    *  PHP 7.4 fix for [#2750](https://github.com/getgrav/grav/issues/2750)
+    * PHP 7.4 fix for [#2750](https://github.com/getgrav/grav/issues/2750)
 
 # v1.6.18
 ## 12/02/2019